+++ elements/geosearch/src/style.js
@@ -20,6 +20,9 @@ const styles = css`
     opacity: 0;
   }
   .results-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
     min-height: 100px;
     width: 332px;
     background: var(--results-bg, #eaf1f5);
@@ -31,6 +34,7 @@ const styles = css`
   input {
     background: var(--input-bg, #c6d4dd);
     color: var(--input-fg, #333);
+    color: #f00;
     height: 48px;
     border-radius: 6px;
     padding: 0 16px;
@@ -47,17 +51,18 @@ const styles = css`
   }
   button {
     height: auto;
-    background: var(--button-bg, #004170);
+    background: var(--button-bg, #fff);
+    color: var(--button-fg, #111);
   }
 
   .geosearch.small button {
     height: var(--button-size, 32px);
     width: var(--button-size, 32px);
-    padding: 6px;
+    padding: calc((var(--button-size, 32px) - 20px) / 2);
   }
 
   .geosearch.small button .icon {
-    min-width: 20px;
+    width: 20px;
     height: 20px;
     transform: translateX(1px);
     background: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' fill='%23fff' viewBox='0 0 24 24'%3E%3Ctitle%3Emagnify%3C/title%3E%3Cpath d='M9.5,3A6.5,6.5 0 0,1 16,9.5C16,11.11 15.41,12.59 14.44,13.73L14.71,14H15.5L20.5,19L19,20.5L14,15.5V14.71L13.73,14.44C12.59,15.41 11.11,16 9.5,16A6.5,6.5 0 0,1 3,9.5A6.5,6.5 0 0,1 9.5,3M9.5,5C7,5 5,7 5,9.5C5,12 7,14 9.5,14C12,14 14,12 14,9.5C14,7 12,5 9.5,5Z' /%3E%3C/svg%3E");