+++ elements/layercontrol/test/cases/general/index.js
@@ -2,6 +2,7 @@ export { default as checkLayerSize } from "./check-layer-size";
 export { default as checkPreOpenLayerTools } from "./check-pre-open-layer-tools";
 export { default as checkPreOpenSection } from "./check-pre-open-section";
 export { default as disableDrag } from "./disable-drag";
+export { default as exclusiveLayers } from "./exclusive-layers";
 export { default as hideLayers } from "./hide-layers";
 export { default as emptyGroup } from "./empty-groups";
 export { default as renderOptionalLayer } from "./render-optional-layer";