+++ elements/storytelling/stories/index.js
@@ -4,3 +4,4 @@ export { default as PrimaryStory } from "./primary"; // StoryTelling with markdo
 export { default as MarkdownAsURLStory } from "./markdown-url"; // StoryTelling with markdown URL
 export { default as MarkdownSlotStory } from "./markdown-slot"; // StoryTelling with markdown from the slot
 export { default as MarkdownEditorStory } from "./markdown-editor"; // StoryTelling with editor
+export { default as MarkdownInitEventStory } from "./markdown-init-event"; // StoryTelling with init event